fix(lint): split oversized files to satisfy the 500-line linecheck cap
The merge pulled reload-on-GET code/tests into four files that grew past
the repo's linecheck threshold (routine_storage.rs, routines/service.rs,
routines/mod_agents_tests.rs, routes/mcp_tests.rs). Split each along an
existing seam, mirroring the repo's established pattern (see #1087):

- routine_storage.rs -> routine_storage_load.rs (load/scan/reload helpers)
- routines/service.rs -> routines/service_update.rs (svc_update)
- routines/mod_agents_tests.rs -> routines/mod_agents_reload_tests.rs
  (the new reload-from-disk svc_list/svc_get tests)
- routes/mcp_tests.rs -> routes/mcp_routine_lifecycle_tests.rs
  (snooze/power-saving/run-listing MCP tool tests)

No behavior change; pure file split to pass `linecheck --max-lines 500`.
